started refactoring the backend routes code for testing- Seperated implementations of functions to backend.src folder
